More renaming of row -> section
