+++ src/platform/packages/private/kbn-grid-layout/grid/grid_section/delete_grid_section_modal.tsx
