+++ src/platform/packages/private/kbn-grid-layout/grid/grid_section/grid_section_drag_preview.tsx
