+++ src/platform/packages/private/kbn-grid-layout/grid/use_grid_layout_events/section/utils.ts
